+++ next-docs/public/examples/avatar/Fallbacks.tsx
@@ -1,0 +1,12 @@
+import { Avatar } from '@heathmont/moon-components';
+
+const Example = () => (
+  <div className="flex justify-around items-center w-full">
+    <Avatar color="gohan.100" backgroundColor="piccolo.100" size="xsmall" />
+    <Avatar color="gohan.100" backgroundColor="piccolo.100" size="small" />
+    <Avatar color="gohan.100" backgroundColor="piccolo.100" size="medium" />
+    <Avatar color="gohan.100" backgroundColor="piccolo.100" size="large" />
+  </div>
+);
+
+export default Example;