+++ next-docs/public/examples/avatar/Image.tsx
@@ -1,0 +1,12 @@
+import { Avatar } from '@heathmont/moon-components';
+
+const Example = () => (
+  <div className="flex justify-around items-center w-full">
+    <Avatar imageUrl="https://www.fillmurray.com/200/200" size="xsmall" />
+    <Avatar imageUrl="https://www.fillmurray.com/200/200" size="small" />
+    <Avatar imageUrl="https://www.fillmurray.com/200/200" size="medium" />
+    <Avatar imageUrl="https://www.fillmurray.com/200/200" size="large" />
+  </div>
+);
+
+export default Example;