+++ next-docs/public/examples/avatar/Letter.tsx
@@ -1,0 +1,35 @@
+import { Avatar } from '@heathmont/moon-components';
+
+const Example = () => (
+  <div className="flex justify-around items-center w-full">
+    <Avatar
+      name="JS"
+      color="gohan.100"
+      backgroundColor="piccolo.100"
+      size="xsmall"
+    />
+
+    <Avatar
+      name="JS"
+      color="gohan.100"
+      backgroundColor="piccolo.100"
+      size="small"
+    />
+
+    <Avatar
+      name="JS"
+      color="gohan.100"
+      backgroundColor="piccolo.100"
+      size="medium"
+    />
+
+    <Avatar
+      name="JS"
+      color="gohan.100"
+      backgroundColor="piccolo.100"
+      size="large"
+    />
+  </div>
+);
+
+export default Example;