+++ packages/assets/src/icons/IconAdd.tsx
@@ -26,7 +26,9 @@ type IconProps = {
   color?: ColorProps,
 };
 const IconAdd =
-  styled(Svg) <
+  styled(Svg).withConfig({
+    shouldForwardProp: prop => !['backgroundColor'].includes(prop),
+  }) <
   IconProps >
   (({ backgroundColor, circleColor, color, theme }) => [
     backgroundColor && {