+++ packages/assets/src/icons/IconCard.tsx
@@ -27,7 +27,9 @@ type IconProps = {
   color?: ColorProps,
 };
 const IconCard =
-  styled(Svg) <
+  styled(Svg).withConfig({
+    shouldForwardProp: prop => !['backgroundColor'].includes(prop),
+  }) <
   IconProps >
   (({ backgroundColor, circleColor, color, theme }) => [
     backgroundColor && {