+++ packages/assets/src/icons/IconCrypto.tsx
@@ -42,7 +42,9 @@ type IconProps = {
   color?: ColorProps,
 };
 const IconCrypto =
-  styled(Svg) <
+  styled(Svg).withConfig({
+    shouldForwardProp: prop => !['backgroundColor'].includes(prop),
+  }) <
   IconProps >
   (({ backgroundColor, circleColor, color, theme }) => [
     backgroundColor && {