+++ packages/assets/src/icons/IconESportsESport.tsx
@@ -39,7 +39,9 @@ type IconProps = {
   color?: ColorProps,
 };
 const IconESportsESport =
-  styled(Svg) <
+  styled(Svg).withConfig({
+    shouldForwardProp: prop => !['backgroundColor'].includes(prop),
+  }) <
   IconProps >
   (({ backgroundColor, circleColor, color, theme }) => [
     backgroundColor && {