+++ packages/assets/src/icons/IconESportsHearthstone.tsx
@@ -38,7 +38,9 @@ type IconProps = {
   color?: ColorProps,
 };
 const IconESportsHearthstone =
-  styled(Svg) <
+  styled(Svg).withConfig({
+    shouldForwardProp: prop => !['backgroundColor'].includes(prop),
+  }) <
   IconProps >
   (({ backgroundColor, circleColor, color, theme }) => [
     backgroundColor && {