+++ packages/assets/src/icons/IconInfoCircle.tsx
@@ -31,7 +31,9 @@ type IconProps = {
   color?: ColorProps,
 };
 const IconInfoCircle =
-  styled(Svg) <
+  styled(Svg).withConfig({
+    shouldForwardProp: prop => !['backgroundColor'].includes(prop),
+  }) <
   IconProps >
   (({ backgroundColor, circleColor, color, theme }) => [
     backgroundColor && {