+++ packages/assets/src/icons/IconLink.tsx
@@ -28,7 +28,9 @@ type IconProps = {
   color?: ColorProps,
 };
 const IconLink =
-  styled(Svg) <
+  styled(Svg).withConfig({
+    shouldForwardProp: prop => !['backgroundColor'].includes(prop),
+  }) <
   IconProps >
   (({ backgroundColor, circleColor, color, theme }) => [
     backgroundColor && {