+++ packages/assets/src/icons/IconSportsFantasy.tsx
@@ -43,7 +43,9 @@ type IconProps = {
   color?: ColorProps,
 };
 const IconSportsFantasy =
-  styled(Svg) <
+  styled(Svg).withConfig({
+    shouldForwardProp: prop => !['backgroundColor'].includes(prop),
+  }) <
   IconProps >
   (({ backgroundColor, circleColor, color, theme }) => [
     backgroundColor && {