+++ packages/assets/src/icons/IconWallet.tsx
@@ -41,7 +41,9 @@ type IconProps = {
   color?: ColorProps,
 };
 const IconWallet =
-  styled(Svg) <
+  styled(Svg).withConfig({
+    shouldForwardProp: prop => !['backgroundColor'].includes(prop),
+  }) <
   IconProps >
   (({ backgroundColor, circleColor, color, theme }) => [
     backgroundColor && {