+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 7.10.1
+
+### Patch Changes
+
+- New docs preview & assets/icons props fix
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@7.10.1
+
 ## 7.10.0
 
 ### Minor Changes